+++ RadarrSharp/Models/LogRecord.cs
@@ -48,7 +48,7 @@ namespace RadarrSharp.Models
         /// <value>
         /// The identifier.
         /// </value>
-        [J("id")] public long Id { get; set; }
+        [J("id")] public int Id { get; set; }
 
         /// <summary>
         /// Gets or sets the exception.
