+++ RadarrSharp/Models/Record.cs
@@ -15,7 +15,7 @@ namespace RadarrSharp.Models
         /// <value>
         /// The movie identifier.
         /// </value>
-        [J("movieId")] public long MovieId { get; set; }
+        [J("movieId")] public int MovieId { get; set; }
 
         /// <summary>
         /// Gets or sets the source title.
@@ -87,6 +87,6 @@ namespace RadarrSharp.Models
         /// <value>
         /// The identifier.
         /// </value>
-        [J("id")] public long Id { get; set; }
+        [J("id")] public int Id { get; set; }
     }
 }
